requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# NLP
spacy>=3.7.0

# Voice
openai-whisper>=20231117
pyttsx3>=2.90

# Audio
pyaudio>=0.2.14
sounddevice>=0.4.6
soundfile>=0.12.1

# Automation
pyautogui>=0.9.54
psutil>=5.9.8
keyboard>=0.13.5

# Database
sqlalchemy>=2.0.0

# Utilities
colorama>=0.4.6
rich>=13.7.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
